+++ casual/casual-network/src/main/java/se/laz/casual/network/inbound/reverse/AutoConnect.java
@@ -1,5 +1,7 @@
 package se.laz.casual.network.inbound.reverse;
 
+import se.laz.casual.api.util.RepeatUntilSuccessTask;
+import se.laz.casual.api.util.StaggeredOptions;
 import se.laz.casual.network.reverse.inbound.ReverseInboundConnectListener;
 import se.laz.casual.network.reverse.inbound.ReverseInboundListener;
 import se.laz.casual.network.reverse.inbound.ReverseInboundServer;
